set up detail page for specific pokemon
