+++ src/App.jsx
@@ -1,5 +1,21 @@
-import {useState, useEffect} from 'react';
+// reach for the toolbox
+// toolbox for routing is 'react-router-dom'
+// npm install react-router-dom
+import {
+  createBrowserRouter, 
+  createRoutesFromElements, 
+  Route,
+  RouterProvider,
+} from 'react-router-dom';
 
+// LAYOUT
+import DefaultLayout from './layout/DefaultLayout';
+
+// PAGES
+import PokemonList from './pages/PokemonList';
+import PokemonDetailView from './pages/PokemonDetailView';
+
+// STYLING
 import './App.css'
 
 // SCSS and BEM
@@ -15,14 +31,18 @@ import './App.css'
 // working with paginated data
 // what is a dependency array and how does it work
 
-import PokemonCard from './components/pokemonCard/PokemonCard';
 
 function App() {
 
-  // this is the state representing our pokemon
-  // well update the state with setPokemon
-  const [pokemon, setPokemon] = useState([]);
-  const [offset, setOffset] = useState(0);
+  const router = createBrowserRouter(
+    createRoutesFromElements (
+      <Route path="/" element={<DefaultLayout/>}>
+        <Route index element={<PokemonList />} />
+        <Route path="/pokemon/:pokemonName" element={<PokemonDetailView />} />
+      </Route>
+  
+    )
+  )
 
   // reach out to some API
   // get some data
@@ -32,61 +52,13 @@ function App() {
 
   // HTTP: GET(read), POST (create), PUT (update), DELETE
   // reach out across the web, to get some pokemon data
-  
-
-  // empty dependency array, runs once when component mounts
-  // useEffect(function() {
-  //   // loadMorePokemon();
-  // }, []); // empty dependency array
-
-  // when offset is updated, run this function
-  useEffect(() => {
-    // do something here
-    const urlForPokemon = `https://pokeapi.co/api/v2/pokemon?limit=10&offset=${offset}`;
-
-    fetch(urlForPokemon)
-      .then(response => response.json())
-      .then(data => {
-        setPokemon(data.results)
-      }).catch(err => {
-        console.error(err.message);
-      })
-  }, [offset])
-
-
-  const loadMorePokemon = () => {
-    // asynchronous  
-    setOffset(offset + 10);
-  }
-
 
   // let setBlackBackground = false;
 
   // iterate through each one and render
+
   return (
-    <div className="pokemon-app">
-      <div className="row">
-        <button 
-          className="pokemon-app__load-more"
-          onClick={loadMorePokemon}
-        > 
-          Load more Pokemon
-        </button>
-        <div className="pokemon-app__page-num">{(offset+10)/10}</div>
-      </div>
-      <div className="promo-card-container">
-        {pokemon.map((singlePokemon, index) => { // index  = 0, 1, 2, 3, 4, 5
-          // if(index % 3 === 0) setBlackBackground = !setBlackBackground;
-          return(
-            <PokemonCard 
-              key={index+1} 
-              singlePokemon={singlePokemon}
-              hasBlackBackground={ Math.floor(index/3) % 2 === 0 ? true : false}
-            /> 
-          )
-        })}
-      </div>
-    </div>
+    <RouterProvider router={router} />
   )
 }
