+++ src/layout/DefaultLayout.jsx
@@ -1,0 +1,17 @@
+
+
+import {
+  Outlet
+} from 'react-router-dom';
+
+const DefaultLayout = () => {
+
+
+  return (
+    <div>
+      <Outlet />
+    </div>
+  )
+}
+
+export default DefaultLayout;
